Modularize test cases
