+++ rhocalc/aims/tests/aims_calc.py
@@ -10,206 +10,6 @@ from rhocalc.aims import aims_calc
 
 TOLERANCE = {"rtol": 1e-5, "atol": 1e-10}
 
-aims_path = "/home/abbott/codes/new_aims/FHIaims/build/aims.230905.scalapack.mpi.x"
-aims_kwargs = {
-    "species_dir": "/home/abbott/codes/new_aims/FHIaims/species_defaults/defaults_2020/tight",
-    "xc": "pbe0",
-    "spin": "none",
-    "relativistic": "none",
-    "charge": 0,
-    "sc_accuracy_rho": 1e-8,
-    "wave_threshold": 1e-8,
-    "elsi_restart": "write 1",
-    # "elsi_restart": "read",
-    # "sc_iter_limit": 0,
-    # "postprocess_anyway": True,
-    # ======================================== ri_fit: what to fit to
-    # "ri_fit_total_density": True,
-    # "ri_fit_ldos": True,
-    # "ri_fit_ildos": True,
-    # ======================================== ri_fit: settings
-    # "ri_fit_ovlp_cutoff_radius": 2.0,
-    # "ri_fit_assume_converged": True,
-    # ======================================== ri_fit: what to write as outputs
-    # "ri_fit_write_nao_coeff_matrix":True,
-    # "ri_fit_write_coeffs": True,
-    # "ri_fit_write_projs": True,
-    # "ri_fit_write_ovlp": True,
-    # "ri_fit_write_ref_field": True,
-    # "ri_fit_write_rebuilt_field": True,
-    # "ri_fit_write_ref_field_cube": True,
-    # "ri_fit_write_rebuilt_field_cube": True,
-    # "ri_fit_sanity_check": True,
-    # ======================================== Keywords we don't want to have to use
-    # "output": ["cube total_density"],
-    # "collect_eigenvectors": True,
-}
-sbatch_kwargs = {
-    "job-name": "checks_scf",
-    "nodes": 1,
-    "time": "01:00:00",
-    "mem-per-cpu": 2000,
-    "partition": "standard",
-}
-
-# Define parameters that are different for each run
-calcs = {
-    0: {
-        "name": "H2O, cluster, serial",
-        "atoms": ase.io.read("systems/water_cluster.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {},
-        "sbatch_kwargs": {"ntasks-per-node": 1},
-    },
-    1: {   
-        "name": "H2O, cluster, parallel, n_tasks = 5",
-        "atoms": ase.io.read("systems/water_cluster.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {},
-        "sbatch_kwargs": {"ntasks-per-node": 5},
-    },
-    2: {
-        "name": "H2O, periodic, 1 kpt, serial",
-        "atoms": ase.io.read("systems/water_periodic.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {"k_grid": [1, 1, 1]},
-        "sbatch_kwargs": {"ntasks-per-node": 1},
-    },
-    3: {  
-        "name": "H2O, periodic, 1 kpt, parallel, n_tasks (5) > n_kpts",
-        "atoms": ase.io.read("systems/water_periodic.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {"k_grid": [1, 1, 1], "collect_eigenvectors": False},
-        "sbatch_kwargs": {"ntasks-per-node": 5},
-    },
-    4: {  
-        "name": "H2O, periodic, 1 kpt, parallel, n_tasks (5) > n_kpts, COLLECT",
-        "atoms": ase.io.read("systems/water_periodic.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {"k_grid": [1, 1, 1], "collect_eigenvectors": True},
-        "sbatch_kwargs": {"ntasks-per-node": 5},
-    },
-    5: {  
-        "name": "H2O, periodic, 4 kpt, parallel, n_tasks (3) < n_kpts",
-        "atoms": ase.io.read("systems/water_periodic.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {"k_grid": [2, 2, 1]},
-        "sbatch_kwargs": {"ntasks-per-node": 3},
-    },
-    6: {  
-        "name": "H2O, periodic, 4 kpt, parallel, n_tasks (3) < n_kpts, COLLECT",
-        "atoms": ase.io.read("systems/water_periodic.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {"k_grid": [2, 2, 1], "collect_eigenvectors": True},
-        "sbatch_kwargs": {"ntasks-per-node": 3},
-    },
-    7: {  
-        "name": "H2O, periodic, 4 kpt, parallel, n_tasks (10) > n_kpts",
-        "atoms": ase.io.read("systems/water_periodic.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {"k_grid": [2, 2, 1]},
-        "sbatch_kwargs": {"ntasks-per-node": 10},
-    },
-    8: {  
-        "name": "H2O, periodic, 4 kpt, parallel, n_tasks (10) > n_kpts, COLLECT",
-        "atoms": ase.io.read("systems/water_periodic.xyz"),
-        "aims_path": aims_path,
-        "aims_kwargs": {"k_grid": [2, 2, 1], "collect_eigenvectors": True},
-        "sbatch_kwargs": {"ntasks-per-node": 10},
-    },
-    # 4: {
-    #     "name": "H2O, periodic, 4 kpt, serial",
-    #     "atoms": ase.io.read("systems/water_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 2, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 1, "time": "03:00:00",},
-    # },
-    # 5: {  
-    #     "name": "H2O, periodic, 4 kpt, parallel, 1 < n_tasks < n_kpts",
-    #     "atoms": ase.io.read("systems/water_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 2, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 3},
-    # },
-    # 6: {  
-    #     "name": "H2O, periodic, 4 kpt, parallel, 1 < n_kpts == n_tasks",
-    #     "atoms": ase.io.read("systems/water_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 2, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 4},
-    # },
-    # 7: {  
-    #     "name": "H2O, periodic, 4 kpt, parallel, 1 < n_kpts < n_tasks",
-    #     "atoms": ase.io.read("systems/water_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 2, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 7},
-    # },
-    # 8: {  
-    #     "name": "H2O, periodic, 4 kpt, parallel, 1 < n_tasks < n_kpts, collect_eigenvectors",
-    #     "atoms": ase.io.read("systems/water_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 2, 2], "collect_eigenvectors": True},
-    #     "sbatch_kwargs": {"ntasks-per-node": 3},
-    # },
-    # 9: {  
-    #     "name": "H2O, periodic, 4 kpt, parallel, 1 < n_kpts == n_tasks, collect_eigenvectors",
-    #     "atoms": ase.io.read("systems/water_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 2, 2], "collect_eigenvectors": True},
-    #     "sbatch_kwargs": {"ntasks-per-node": 4},
-    # },
-    # 10: {  
-    #     "name": "H2O, periodic, 4 kpt, parallel, 1 < n_kpts < n_tasks, collect_eigenvectors",
-    #     "atoms": ase.io.read("systems/water_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 2, 2], "collect_eigenvectors": True},
-    #     "sbatch_kwargs": {"ntasks-per-node": 7},
-    # },
-    # 11: {  
-    #     "name": "H2O, periodic, 4 kpt, serial, larger cell",
-    #     "atoms": ase.io.read("systems/water_periodic_larger_cell.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 2, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 1},
-    # },
-    # 12: {  
-    #     "name": "H2O, periodic, 1 kpt, parallel, n_tasks > n_kpts, larger cell",
-    #     "atoms": ase.io.read("systems/water_periodic_larger_cell.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 1, 1]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 10},
-    # },
-    # 13: {  
-    #     "name": "H2O, periodic, 2 kpt, parallel, 1 < n_tasks == n_kpts",
-    #     "atoms": ase.io.read("systems/water_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 1, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 2},
-    # },
-    # 14: {  
-    #     "name": "H2O, periodic, 2 kpt, parallel, 1 < n_tasks == n_kpts, larger cell",
-    #     "atoms": ase.io.read("systems/water_periodic_larger_cell.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 1, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 2},
-    # },
-    # 15: {  
-    #     "name": "Si, periodic, 2 kpt, parallel, 1 < n_tasks == n_kpts",
-    #     "atoms": ase.io.read("systems/si_periodic.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 1, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 2},
-    # },
-    # 16: {  
-    #     "name": "Si, periodic, 2 kpt, parallel, 1 < n_tasks == n_kpts, larger cell",
-    #     "atoms": ase.io.read("systems/si_periodic_larger_cell.xyz"),
-    #     "aims_path": aims_path,
-    #     "aims_kwargs": {"k_grid": [1, 1, 2]},
-    #     "sbatch_kwargs": {"ntasks-per-node": 2},
-    # },
-}
-
 
 def run_scf(aims_kwargs: dict, sbatch_kwargs: dict, calcs: dict):
     """
@@ -331,47 +131,6 @@ def run_ri(aims_kwargs: dict, sbatch_kwargs: dict, calcs: dict, use_restart: boo
 # ======================================== 
 # ========= Test functions ===============
 # ========================================
-
-
-def run_tests(calcs: dict):
-    """
-    Runs a series of tests on AIMS outputs fpr each of the calculations in
-    `calcs`. Each test returns the quantities that failed comparison tests. This
-    function returns a list of
-    """
-    failed = []
-
-    # # Get integration to the formal number of electrons
-    # failed.append(total_densities_integrate_to_n_electrons(calcs))
-
-    # # np.allclose of weighted sum of KSO coeff matrices vs density matrix
-    # failed.append(coeff_matrices_sum_to_density_matrix(calcs, return_all_calcs=True))
-
-    # MAE of total density between a) that built from densmat and b) that imported from physics
-    failed.append(density_from_densmat_equals_density_from_physics(calcs))
-
-    # # MAE of total density built from RI coefficients vs that built from NAO densmat
-    # failed.append(density_from_ri_equals_density_from_densmat(calcs))
-
-    # # MAE of total density between a) built from sum of KSOs and b) built from densmat
-    # failed.append(ksos_from_coeffmats_sum_to_density_from_densmat(calcs))
-
-    # # MAE of total density between a) built from sum of KSOs and b) imported from physics
-    # failed.append(ksos_from_coeffmats_sum_to_density_from_physics(calcs))
-
-    # # MAE of total density between a) built from sum of KSOs and b) built from densmat
-    # failed.append(ksos_from_ri_sum_to_density_from_densmat(calcs))
-
-    # # MAE of total density between a) built from sum of KSOs and b) imported from physics
-    # failed.append(ksos_from_ri_sum_to_density_from_physics(calcs))
-
-    # # np.allclose of RI coeffs: a) weighted sum of KSOs vs total density from densmat
-    # failed.append(ri_coeffs_for_ksos_sum_to_total_density_from_densmat(calcs))
-
-    # Test the the RI overlap matric is symmetric
-    # failed.append(overlap_is_symmetric(calcs))
-
-    return failed
 
 
 def total_densities_integrate_to_n_electrons(calcs):
